requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
httpx
sqlalchemy
aiosqlite
python-dotenv
asyncpg
stripe
alembic
psycopg2-binary
PyJWT
redis[async]
boto3
